+++ packages/stream_chat/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `PaginationParams.createdAtBefore` for message pagination.
 - Added `PaginationParams.createdAtAround` for message pagination.
 - Added support for `channel.disabled`, `channel.hidden` and `channel.truncatedAt` in `Channel`.
+- Added support for `channel.membership` and `channel.membershipStream` in `Channel`.
 
 🔄 Changed